[WIP]
- socket on in DmChatUI
