+++ frontend/src/components/Chat/DmList.svelte
@@ -90,38 +90,6 @@
         }
     }
 
-    /*
-    * from dmPageFile
-    */
-    import { socketStore } from '$lib/webSocketConnection_chat';
-	import type { Socket } from 'socket.io-client';
-	import { onDestroy, onMount } from 'svelte';
-	import type { DmChatIF } from '$lib/interface';
-
-	let socket: Socket;
-    let dmChatData : DmChatIF;
-    
-    const unsubscribe = socketStore.subscribe((_socket: Socket) => {
-        socket = _socket;
-	});
-    
-    onDestroy(unsubscribe);
-
-    function sendDm(opponent : string)
-    {
-        dmStoreData[opponent]._dmChatStore.push(dmChatData);
-        localStorage.setItem(DM_KEY, JSON.stringify(dmStoreData));
-        if (dmChatData._msg.length && dmChatData._msg != '\n')
-            socket.emit('dm-chat', dmChatData);
-    }
-
-    function sendDm(opponent : string)
-    {
-        dmStoreData[opponent]._dmChatStore.push(dmChatData);
-        localStorage.setItem(DM_KEY, JSON.stringify(dmStoreData));
-        if (dmChatData._msg.length && dmChatData._msg != '\n')
-            socket.emit('dm-chat', dmChatData);
-    }
 
 </script>
